Compare MeshInstance::frameCounter as the unsigned byte the binary does
UpdateInterpolatedFields' second guard half tests the instance's frame
stamp against MeshInstance::sFrameCounter. The binary compares them as
raw bytes - mov al, [ebp+0D8h] then cmp al against the uchar static, at
0x007DECB9 - but the lane was typed std::int8_t here, so the two sides
promoted to int with different signs for every value above 127.

sFrameCounter is a wrapping uint8_t bumped once per frame, so half of
each 256-frame cycle compared unequal when the binary compares equal,
and the 'already refreshed this frame' short-circuit never fired across
that half. Retype the lane unsigned and drop the four int8 casts that
were feeding it (plus the one in UserUnit); the ctor's -1 becomes the
same 0xFF byte.

+++ src/sdk/moho/unit/core/UserUnit.cpp
@@ -4301,7 +4301,7 @@ void UserUnit::UpdateVisibility()
     meshInstance->isHidden = ((mIntelStateFlags & 0x10u) == 0u) ? 1u : 0u;
     meshInstance->isLocked = ((mIntelStateFlags & 0x08u) == 0u) ? 1u : 0u;
     if (meshInstance->isLocked == 0u) {
-      meshInstance->frameCounter = static_cast<std::int8_t>(MeshInstance::sFrameCounter);
+      meshInstance->frameCounter = MeshInstance::sFrameCounter;
       meshInstance->currInterpolant = -1.0f;
     }
     break;
